requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-log==0.4.0
colorama==0.4.6
deprecation==2.1.0
docutils==0.22.2
dotty-dict==1.3.1
fastapi==0.104.1
gitdb==4.0.12
GitPython==3.1.45
gotrue==1.3.1
h11==0.14.0
httpcore==0.16.3
httptools==0.6.4
httpx==0.23.3
idna==3.10
importlib_metadata==8.7.0
invoke==1.7.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.6.0
more-itertools==10.8.0
nh3==0.3.0
packaging==25.0
pkginfo==1.12.1.2
postgrest==0.10.7
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-gitlab==3.15.0
python-semantic-release==7.33.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
railway==0.0.4
readme_renderer==44.0
realtime==1.0.6
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.27.0
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tomlkit==0.13.3
tqdm==4.67.1
twine==3.8.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==12.0
wheel==0.45.1
zipp==3.23.0
